requirements.txt
streamlit==1.28.0
google-generativeai==0.8.3
python-dotenv==1.0.0
imap-tools==1.0.0
python-dateutil==2.8.2
pandas==2.0.3
plotly==5.15.0
